+++ spec/lib/graphql_rails/controller/action_hooks_runner_spec.rb
@@ -23,12 +23,19 @@ module GraphqlRails
         end
       end
 
-      subject(:runner) { described_class.new(action_name: action_name, controller: controller) }
+      subject(:runner) do
+        described_class.new(
+          action_name: action_name,
+          controller: controller,
+          graphql_request: graphql_request
+        )
+      end
 
       let(:action_name) { :some_action }
       let(:controller) { double('Controller') } # rubocop:disable RSpec/VerifiedDoubles
       let(:controller_class) { class_double(Controller) }
       let(:controller_configuration) { Controller::Configuration.new(controller) }
+      let(:graphql_request) { GraphqlRails::Controller::Request.new(nil, {}, {}) }
 
       before do
         allow(controller).to receive(:class).and_return(controller_class)